Rename test package to run for consistency
