+++ go.mod
@@ -8,6 +8,7 @@ require (
 	github.com/axw/gocov v1.0.0
 	github.com/bmatcuk/doublestar/v2 v2.0.4
 	github.com/breml/logstash-config v0.4.0
+	github.com/chzyer/test v0.0.0-20180213035817-a1ea475d72b1
 	github.com/go-playground/overalls v0.0.0-20191218162659-7df9f728c018
 	github.com/golang/protobuf v1.4.2
 	github.com/hashicorp/packer v1.4.4
