+++ internal/app/daemon/run/run.go
@@ -1,4 +1,4 @@
-package test
+package run
 
 import (
 	"context"
