+++ internal/app/daemon_run.go
@@ -4,7 +4,7 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 
-	"github.com/magnusbaeck/logstash-filter-verifier/v2/internal/app/daemon/test"
+	"github.com/magnusbaeck/logstash-filter-verifier/v2/internal/app/daemon/run"
 	"github.com/magnusbaeck/logstash-filter-verifier/v2/internal/logging"
 )
 
@@ -29,7 +29,7 @@ func runDaemonRun(_ *cobra.Command, args []string) error {
 	// TODO: Remove this
 	// logging.SetLevel(oplogging.INFO)
 
-	t, err := test.New(viper.GetString("socket"), viper.Get("logger").(logging.Logger), viper.GetString("pipeline"), viper.GetString("pipeline-base"), viper.GetString("testcase-dir"))
+	t, err := run.New(viper.GetString("socket"), viper.Get("logger").(logging.Logger), viper.GetString("pipeline"), viper.GetString("pipeline-base"), viper.GetString("testcase-dir"))
 	if err != nil {
 		return err
 	}
